rename minerva-stack to minervaStack
